+++ src/observables/dnd.ts
@@ -2,7 +2,8 @@ import { fromEvent, Observable } from 'rxjs';
 import { filter, map, skipWhile, switchMap, tap } from 'rxjs/operators';
 
 import { Offset, OffsetNumbers } from '../types';
-import { scaleOfElement } from '../utils/dom';
+import { offsetOfElement, scaleOfElement } from '../utils/dom';
+import { snapObjectValues } from '../utils/misc';
 import { allMoveEnd$, allMoveStart$, allMoveUpdate$ } from './allMove';
 import {
     documentMouseMove$,
@@ -11,9 +12,18 @@ import {
 } from './misc';
 
 interface DndObservableOptions {
+    // HTML element used as a basis for all calculations.
     element: HTMLElement;
+
+    // Elements whose observables belong to the same
+    // group will respond to each others movements.
     group: string;
+
+    // HTML element used as a target for mouse interactions.
     handle: HTMLElement;
+
+    // Round values to an interval of this number.
+    snapTo?: number;
 }
 
 /**
@@ -24,6 +34,7 @@ export const createDndObservable = ({
     element,
     group,
     handle,
+    snapTo,
 }: DndObservableOptions): Observable<Offset> => {
     const mouseDown$ = fromEvent(handle, 'mousedown');
 
@@ -34,12 +45,15 @@ export const createDndObservable = ({
             e.preventDefault();
             e.stopPropagation();
 
+            const originalOffset = offsetOfElement(element);
+
             // Account for CSS transform scale
             const scale = scaleOfElement(element);
 
             const move$ = documentMouseMove$.pipe(
                 map(changeFromPointToMouseEvent(e.clientX, e.clientY, scale)),
                 skipWhile(hasntMovedFivePixels),
+                map(adjustForSnap(originalOffset, snapTo)),
                 tap(notifyListeners(group))
             );
 
@@ -100,6 +114,36 @@ const changeFromPointToMouseEvent = (
  */
 const hasntMovedFivePixels = (change: OffsetNumbers) =>
     Math.sqrt(change.top * change.top + change.left * change.left) < 5;
+
+/**
+ * Adjusts the offset changes so that the element that is being moved
+ * ends up in a position that is snapped to the given number.
+ *
+ * NOTE: If you have grouped elements (via the `group` param),
+ * only the element that is being interacted with via mouse
+ * will snap, with all other elements simply following along.
+ */
+const adjustForSnap = (original: OffsetNumbers, snapTo?: number) => (
+    change: OffsetNumbers
+): OffsetNumbers => {
+    if (!snapTo) {
+        return change;
+    }
+
+    const unsnappedOffset: OffsetNumbers = {
+        left: original.left + change.left,
+        top: original.top + change.top,
+    };
+
+    const snapped = snapObjectValues(snapTo)(unsnappedOffset);
+
+    const adjustedChange = {
+        left: snapped.left - original.left,
+        top: snapped.top - original.top,
+    };
+
+    return adjustedChange;
+};
 
 /**
  * Notify allMove observers that a change in mouse position