+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2019-02-11
+### Changed
+- When using `snapTo` with grouped elements, the element that is being
+  interacted with via mouse will snap, with all other elements simply
+  following along.
+
 ## [0.4.0] - 2019-02-08
 ### Added
 - Keyboard-based movement can be disabled
@@ -34,3 +40,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-12-02
 ### Added
 - Initial public release
+